mv gamma epi to dedicated submodule
