+++ episimlab/setup/greek/gamma.py
@@ -3,15 +3,12 @@ import xarray as xr
 import numpy as np
 import logging
 
-from ...seir.base import BaseSEIR
-from .base import BaseSetupEpi
 from ...utils.rng import get_rng
 
 
 @xs.process
-class SetupDefaultGamma(BaseSetupEpi):
-    """
-    """
+class SetupDefaultGamma:
+    """Provide default values for `gamma` for compartments Ia, Iy, and Ih."""
 
     gamma = xs.global_ref('gamma', intent='out')
 
@@ -30,41 +27,37 @@ class SetupDefaultGamma(BaseSetupEpi):
 
 
 @xs.process
-class SetupStaticGamma(SetupDefaultGamma):
-    """Given a length 3 iterable input `tri_exposed_para`, calculate gamma
-    after sampling once from this triangular distibution, at the beginning of
-    the simulation.
+class SetupGammaIh:
+    """Draws `gamma` for compartment Ih from a triangular distribution
+    defined by 3-length array `tri_h2r`.
     """
+    gamma_Ih = xs.global_ref('gamma_Ih', intent='out')
     tri_h2r = xs.variable(dims=('value'), static=True, intent='in')
-    tri_y2r_para = xs.variable(dims=('value'), static=True, intent='in')
-    stochastic = xs.foreign(BaseSEIR, 'stochastic', intent='in')
-    seed_state = xs.foreign(BaseSEIR, 'seed_state', intent='in')
+    stochastic = xs.global_ref('stochastic', intent='in')
+    seed_state = xs.global_ref('seed_state', intent='in')
 
     def get_gamma(self) -> xr.DataArray:
-        dims = ["compartment"]
-        da = xr.DataArray(
-            data=0.,
-            dims=dims,
-            coords={k: self.counts_coords[k] for k in dims}
-        )
-        da.loc[dict(compartment=['Ia', 'Iy', 'Ih'])] = [
-            self.get_gamma_a(),
-            self.get_gamma_y(),
-            self.get_gamma_h(),
-        ]
-        return da
-
-    def get_gamma_h(self) -> float:
-        """Sample from triangular distributions if stochastic, or return the
-        mean if deterministic.
-        """
         if self.stochastic is True:
             rng = get_rng(seed=self.seed_state)
             return 1 / rng.triangular(*self.tri_h2r)
         else:
             return 1 / np.mean(self.tri_h2r)
+    
+    def initialize(self):
+        self.gamma_Ih = self.get_gamma()
 
-    def get_gamma_y(self) -> float:
+
+@xs.process
+class SetupGammaIa:
+    """Draws `gamma` for compartments Ia from a triangular distribution
+    defined by 3-length array `tri_y2r_para`.
+    """
+    gamma_Ia = xs.global_ref('gamma_Ia', intent='out')
+    tri_y2r_para = xs.variable(dims=('value'), static=True, intent='in')
+    stochastic = xs.global_ref('stochastic', intent='in')
+    seed_state = xs.global_ref('seed_state', intent='in')
+
+    def get_gamma_Ia(self) -> float:
         """Sample from triangular distributions if stochastic, or return the
         mean if deterministic.
         """
@@ -74,18 +67,8 @@ class SetupStaticGamma(SetupDefaultGamma):
         else:
             return 1 / np.mean(self.tri_y2r_para)
 
-    def get_gamma_a(self) -> float:
-        """Sample from triangular distributions if stochastic, or return the
-        mean if deterministic.
+    def initialize(self):
+        """Note that because `seed_state` is same for both rng, gamma_Ia
+        and gamma_Iy are exactly the same at each timestep.
         """
-        return self.get_gamma_y()
-
-
-@xs.process
-class SetupDynamicGamma(SetupStaticGamma):
-    """Like SetupStaticGamma, but the triangular distibution is sampled
-    to calculate gamma at every step.
-    """
-
-    def run_step(self):
-        self.gamma = self.get_gamma()
+        self.gamma_Ia = self.get_gamma_Ia()
